+++ spotlight-client/package.json
@@ -33,6 +33,7 @@
     "@types/react": "^16.9.0",
     "@types/react-dom": "^16.9.0",
     "@types/react-measure": "^2.0.6",
+    "@types/react-portal": "^4.0.2",
     "@types/react-simple-maps": "^1.0.3",
     "@types/styled-components": "^5.1.5",
     "@types/topojson": "^3.2.2",
@@ -68,6 +69,7 @@
     "react-intersection-observer": "^8.31.0",
     "react-is": "^16.13.1",
     "react-measure": "^2.5.2",
+    "react-portal": "^4.2.1",
     "react-scripts": "3.4.3",
     "react-simple-maps": "^2.3.0",
     "react-spring": "^8.0.27",
