+++ spotlight-client/src/UiLibrary/Dropdown/DropdownMultiple.test.tsx
@@ -17,8 +17,13 @@
 
 import { render, fireEvent, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
+import useBreakpoint from "@w11r/use-breakpoint";
 import React from "react";
 import DropdownMultiple from "./DropdownMultiple";
+
+jest.mock("@w11r/use-breakpoint");
+
+const useBreakpointMock = useBreakpoint as jest.Mock;
 
 const testLabel = "Test Label";
 const mockOnChange = jest.fn();
@@ -28,10 +33,6 @@ const testOptions = [
   { id: "2", label: "option two" },
   { id: "3", label: "option three" },
 ];
-
-beforeEach(() => {
-  jest.resetAllMocks();
-});
 
 let rendered: ReturnType<typeof render>;
 
@@ -47,100 +48,121 @@ beforeEach(() => {
   );
 });
 
-test("selects multiple", () => {
-  const menuButton = screen.getByRole("button", {
-    name: `${testLabel} ${testOptions[0].label} and 2 others`,
-  });
-
-  fireEvent.click(menuButton);
-
-  testOptions.forEach((opt) => {
-    expect(
-      screen.getByRole("option", { name: opt.label, selected: true })
-    ).toBeVisible();
-  });
+afterEach(() => {
+  jest.resetAllMocks();
 });
 
-test("menu stays open after selection", async () => {
-  const menuButton = screen.getByRole("button", {
-    name: `${testLabel} ${testOptions[0].label} and 2 others`,
-  });
+describe.each([["mobile", true], ["desktop"]])(
+  "on %s:",
+  (...[label, isMobile]) => {
+    beforeEach(() => {
+      useBreakpointMock.mockReturnValue(isMobile);
+    });
 
-  fireEvent.click(menuButton);
+    test("selects multiple", () => {
+      const menuButton = screen.getByRole("button", {
+        name: `${testLabel} ${testOptions[0].label} and 2 others`,
+      });
 
-  fireEvent.click(screen.getByRole("option", { name: testOptions[0].label }));
-  expect(mockOnChange).toHaveBeenLastCalledWith([
-    testOptions[1].id,
-    testOptions[2].id,
-  ]);
-  // update the controlled value to simulate real use
-  rendered.rerender(
-    <DropdownMultiple
-      label={testLabel}
-      options={testOptions}
-      onChange={mockOnChange}
-      selectedIds={["2", "3"]}
-    />
-  );
+      fireEvent.click(menuButton);
 
-  // let's give the animation some time to run;
-  // its duration is not entirely predictable but this should be enough time
-  await new Promise((resolve) => setTimeout(resolve, 750));
+      testOptions.forEach((opt) => {
+        expect(
+          screen.getByRole("option", { name: opt.label, selected: true })
+        ).toBeVisible();
+      });
+    });
 
-  const option2 = screen.getByRole("option", { name: testOptions[1].label });
-  expect(option2).toBeVisible();
+    test("menu stays open after selection", async () => {
+      const menuButton = screen.getByRole("button", {
+        name: `${testLabel} ${testOptions[0].label} and 2 others`,
+      });
 
-  fireEvent.click(option2);
-  expect(mockOnChange).toHaveBeenLastCalledWith([testOptions[2].id]);
-});
+      fireEvent.click(menuButton);
 
-test("select all", () => {
-  const menuButton = screen.getByRole("button", {
-    name: `${testLabel} ${testOptions[0].label} and 2 others`,
-  });
+      fireEvent.click(
+        screen.getByRole("option", { name: testOptions[0].label })
+      );
+      expect(mockOnChange).toHaveBeenLastCalledWith([
+        testOptions[1].id,
+        testOptions[2].id,
+      ]);
+      // update the controlled value to simulate real use
+      rendered.rerender(
+        <DropdownMultiple
+          label={testLabel}
+          options={testOptions}
+          onChange={mockOnChange}
+          selectedIds={["2", "3"]}
+        />
+      );
 
-  fireEvent.click(menuButton);
+      // let's give the animation some time to run;
+      // its duration is not entirely predictable but this should be enough time
+      await new Promise((resolve) => setTimeout(resolve, 750));
 
-  const selectAll = screen.getByRole("option", { name: "Deselect all" });
-  expect(selectAll).toBeVisible();
+      const option2 = screen.getByRole("option", {
+        name: testOptions[1].label,
+      });
+      expect(option2).toBeVisible();
 
-  fireEvent.click(selectAll);
-  expect(mockOnChange).toHaveBeenLastCalledWith([]);
+      fireEvent.click(option2);
+      expect(mockOnChange).toHaveBeenLastCalledWith([testOptions[2].id]);
+    });
 
-  // update the controlled value to simulate real use; none are selected
-  rendered.rerender(
-    <DropdownMultiple
-      label={testLabel}
-      options={testOptions}
-      onChange={mockOnChange}
-      selectedIds={[]}
-    />
-  );
+    test("select all", () => {
+      const menuButton = screen.getByRole("button", {
+        name: `${testLabel} ${testOptions[0].label} and 2 others`,
+      });
 
-  expect(selectAll).toHaveTextContent("Select all");
-  expect(
-    screen.queryByRole("option", { selected: true })
-  ).not.toBeInTheDocument();
+      fireEvent.click(menuButton);
 
-  fireEvent.click(selectAll);
-  expect(mockOnChange).toHaveBeenLastCalledWith(
-    testOptions.map(({ id }) => id)
-  );
-});
+      const selectAll = screen.getByRole("option", { name: "Deselect all" });
+      expect(selectAll).toBeVisible();
 
-test("callback for highlighted item", async () => {
-  const menuButton = screen.getByRole("button", {
-    name: `${testLabel} ${testOptions[0].label} and 2 others`,
-  });
+      fireEvent.click(selectAll);
+      expect(mockOnChange).toHaveBeenLastCalledWith([]);
 
-  userEvent.click(menuButton);
+      // update the controlled value to simulate real use; none are selected
+      rendered.rerender(
+        <DropdownMultiple
+          label={testLabel}
+          options={testOptions}
+          onChange={mockOnChange}
+          selectedIds={[]}
+        />
+      );
 
-  userEvent.hover(await screen.findByRole("option", { name: "Deselect all" }));
-  // it does get called but indicates that nothing should be highlighted
-  // (e.g. to clear an existing one)
-  expect(mockOnHighlight).toHaveBeenLastCalledWith();
+      expect(selectAll).toHaveTextContent("Select all");
+      expect(
+        screen.queryByRole("option", { selected: true })
+      ).not.toBeInTheDocument();
 
-  userEvent.hover(screen.getByRole("option", { name: testOptions[0].label }));
+      fireEvent.click(selectAll);
+      expect(mockOnChange).toHaveBeenLastCalledWith(
+        testOptions.map(({ id }) => id)
+      );
+    });
 
-  expect(mockOnHighlight).toHaveBeenLastCalledWith(testOptions[0].id);
-});
+    test("callback for highlighted item", async () => {
+      const menuButton = screen.getByRole("button", {
+        name: `${testLabel} ${testOptions[0].label} and 2 others`,
+      });
+
+      userEvent.click(menuButton);
+
+      userEvent.hover(
+        await screen.findByRole("option", { name: "Deselect all" })
+      );
+      // it does get called but indicates that nothing should be highlighted
+      // (e.g. to clear an existing one)
+      expect(mockOnHighlight).toHaveBeenLastCalledWith();
+
+      userEvent.hover(
+        screen.getByRole("option", { name: testOptions[0].label })
+      );
+
+      expect(mockOnHighlight).toHaveBeenLastCalledWith(testOptions[0].id);
+    });
+  }
+);
